Added form country dropdown for guest and host -ref SOSUA-77
